// src/app/Zwibbler/zwibbler2.d.ts
export interface ZwibblerClass {
    /** Contains references to all of the ZwibblerContexts in use. */
    "instances": ZwibblerContext[]

    /** Adds a TTF or OTF font for use in PDF files.
     * @param url The url of the font
     * @return aApromise resolving to the name of the font.
     */
    addFont(url: string): PromiseLike<string>;

    /** Attaches a scope to an HTML element, executing all directives.
     * @param id A CSS selector or reference to the element.
     * @param scope The scope to attach.
     */
    attach(id: string | Element, scope: any): MainScope;

    /** Given an SVG Path (as would be used in the d attribute of a path element),
     * returns an array of commands suitable for use in a Zwibbler PathNode.
     */
    commandsFromSVGPath(d: string): number[];

    /** Defines a component for use in the Zwibbler framework, or
     * to be used with in createHTMLNode.
     * 
     * @param name The name of the component
     * @param component Defines the component
     */
    component(name: string, component: ComponentWithProperties): void;

    /** Defines a controller function to be used with 
     *  &lt;zwibbler z-controller=...
     * 
     * @param name The name of the controller that will be specified in z-controller
     * @param fn The controller function to be called.
     */
    controller(name: string, fn: (scope: MainScope) => void): void;

    /** Creates a zwibbler instance, filling the given element with the drawing area.
     * @param id The CSS selector or reference to the element that will contain the drawing area.
     * @param options Optional configuration settings. See Zwibbler configuration settings.
     */
    create(id: string | Element, options?: Properties): ZwibblerContext;

    /** Creates a component previously defined with Zwibbler.component().
     * @param parentScope The parent scope. Any properties defined in the component will be made available to the component.
     * @param name The name of the component, previously registered with Zwibbler.component
     * @param attributes A mapping from property name to expressions. The expression, evaluated on the current scope, will be assigned to the property on the child scope.
     * 
     * @return both the HTML element and the child scope.
     */
    createComponent(parentScope: any, name: string, attributes?: { [name: string]: string }, childScope?: any): { "node": HTMLElement, "scope": Scope }

    /** Destroys a component created with createComponent, removing any event listeners.*/
    destroyComponent(el: HTMLElement): void

    /** Detach an element previously attached with attach, destroying components and detaching event listeners. */
    detach(el: HTMLElement): void;

    directive(name: string, fn: (d: DirectiveContext) => void): void;

    enableConsoleLogging(): void;

    formatText(ctx: ICanvasContext, o: TextOptions): FormattedText;

    /** Inject some CSS into the web page. The CSS is injected once time, even if called multiple times
     * for the same cssText. It can be removed by calling Zwibbler.destroyAll()
     */
    injectStyle(cssText: string): void;

    /** 
     <ul>
     <li> Destroys an open instances of Zwibbler.
     <li> Removes any styles that have been injected into the web page.
     <li> Forgets any custom nodes or buttons.
     <li> Forgets fonts added by addFont()
     </ul>
     */
    destroyAll(): void;

    /** Converts an rgba object to a string colour.
     * The returned colour is either of the form #000000 or rgba(0,0,0,0.0)
     * if it has transparency.
    */
    makeColour(clr: EColour): string;

    /* Converts a colour string to r, g, b, and a values.
       The values are between 0 and 1.0. */
    parseColour(str: string): EColour;

    /**
      Alters a colour to have an opacity value. The colour may be in # or rgb, rgba format.

      @param clr The colour, in one of these formats: #ffffff, #fff, rgba(255, 255, 255, 1.0), rgb(255, 255, 255)
      @param opacity Between 0 (fully transparent) and 1 (fully opaque)

      @returns colour in rgba or #ffffff format.
     */
    setColourOpacity(clr: string, opacity: number): string;

    setHashValue(name: string, value: string): void;

    /** Loads an image as a promise */
    loadImage(url: string): Promise<HTMLImageElement>;
}
// ...
export type Properties = IAny;
interface IAny {
    [key: string]: any;
}

export interface MainScope extends IAny {
    "ctx": ZwibblerContext;
    "showPopup"(name: string, args?: ShowPopupArgs): void;
    "hidePopup"(name: string): void;
}

export interface Scope extends MainScope {
    "ctx": ZwibblerContext;
    "element": HTMLElement;
    "id": string;
    "parent": Scope;
    "children": Scope[];
    "ze": null;
    "type": string;
    "props": Properties;
}

type ControllerFn = (scope: Scope) => void;

export interface ComponentWithProperties extends Component {
    "propertyPanel"?: string;
    "propertyPanelController"?: ControllerFn;
    "defaults"?: any;
    "controller"?: ControllerFn | null;
    "draw"?: (scope: any, context: ICanvasContext) => void;
}

export interface Component {
    "properties"?: string[];
    "style"?: string;
    "template"?: string | HTMLElement;
    "controller"?: ((scope: any, info: DirectiveContext) => void) | null;
}

export interface DirectiveContext {
    "scope": any;
    "element": HTMLElement;
    "name": string;
    "value": string;
    "emit"(eventName: string, arg: any): void;
    "listen"(eventName: string, fn: (this: HTMLElement, e: any) => void): void;
    "watch"(expr: string, changeFn: (newValue: any) => void): void;
    "eval"(expr: string): any;
    "destructor"(fn: () => void): void;
}

export type NodeID = string;
export type NodeIDs = NodeID | NodeID[];

export interface ToolbarItem {
    "toolName"?: string;
    "onclick"?: (this: HTMLElement, ctx: ZwibblerContext, e: Event) => void;
    "title"?: string;
    "background"?: string;
    "image"?: string;
    "html"?: string;
}

export interface ShowPopupArgs {
    "x"?: number
    "y"?: number
    "refElement"?: HTMLElement
    "showPosition"?: string | null
    "clickToDismiss"?: string | null
    "parent"?: HTMLElement
    "onhide"?: () => void;

    // Overlay colour (default: "transparent")
    "overlay"?: string
}


interface ExternalPoint {
    "x": number;
    "y": number;
}

interface ExternalRect extends ExternalPoint {
    "width": number;
    "height": number;
}

export interface ICanvasDrawContext {
    arc(x: number, y: number, radius: number, sa: number, ea: number, counterclockwise?: boolean): void;
    arcTo(x1: number, y1: number, x2: number, y2: number, radius: number): void;
    beginPath(): void;
    closePath(): void;
    moveTo(x: number, y: number): void;
    lineTo(x: number, y: number): void;
    quadraticCurveTo(x1: number, y1: number, x2: number, y2: number): void;
    bezierCurveTo(x1: number, y1: number, x2: number, y2: number, x3: number, y3: number): void;
    rect(a: number, b: number, c: number, d: number): void;
}
export declare type HTMLImageSource = HTMLImageElement | HTMLCanvasElement;
export declare type ICanvasPattern = {};
export declare type ICanvasGradient = {
    addColorStop(offset: number, color: string): void;
};

/** 
 * The ICanvasContext is a subset of behaviour of HTML canvas.
 * Zwibbler guarantees to implement this on SVG, PDF, as well as on-screen.
 */
export interface ICanvasContext extends ICanvasDrawContext {
    setTransform(a: number, b: number, c: number, d: number, e: number, f: number): void;
    transform(a: number, b: number, c: number, d: number, e: number, f: number): void;
    translate(x: number, y: number): void;
    rotate(a: number): void;
    scale(x: number, y: number): void;
    stroke(): void;
    fill(): void;
    fillText(text: string, x: number, y: number): void;
    strokeText(text: string, x: number, y: number): void;
    fillRect(x: number, y: number, width: number, height: number): void;
    strokeRect(x: number, y: number, width: number, height: number): void;
    save(): void;
    restore(): void;
    measureText(s: string): {
        width: number;
    };
    drawImage(image: HTMLImageSource, dx: number, dy: number): void;
    drawImage(image: HTMLImageSource, dx: number, dy: number, dWidth: number, dHeight: number): void;
    drawImage(image: HTMLImageSource, sx: number, sy: number, sWidth: number, sHeight: number, dx: number, dy: number, dWidth: number, dHeight: number): void;
    createLinearGradient(x0: number, y0: number, x1: number, y1: number): ICanvasGradient;
    clip(): void;
    textBaseline: string;
    fillStyle: string | ICanvasPattern | ICanvasGradient;
    strokeStyle: string | ICanvasPattern | ICanvasGradient;
    globalCompositeOperation: string;
    globalAlpha: number;
    lineWidth: number;
    ZwibblerBackgroundPattern?: CanvasPattern | null;
    font: string;
    shadowOffsetX: number;
    shadowOffsetY: number;
    shadowBlur: number;
    shadowColor: string;
    lineJoin: string;
    lineCap: string;
    lineDashOffset: number;
    miterLimit: number;

    /** In SVG, starts a new &lt;g> tag with the given attributes. */
    pushGroup?(attributes: {
        [key: string]: string;
    }): void;

    /** In SVG, closes a &lt;g> tag previously opened with pushGroup. */
    popGroup?(): void;

    createPattern(src: HTMLImageSource, how: string): CanvasPattern | null;
}
// ...
export interface TextOptions {
    "text": string;
    "width"?: number;
    "height"?: number;
    "font"?: string;
    "valign"?: string;
    "halign"?: string,
    "textDecoration"?: string;
}

export interface FormattedText {
    "draw"(ctx: ICanvasContext, x: number, y: number): void;
    "width": number
    "height": number
}

export interface Decoration {
    appliesTo?(node: NodeID, ctx: ZwibblerContext): boolean;
    xoffset?: number;
    yoffset?: number;
    x?: number;
    y?: number;
    width?: number;
    onclick?(node: NodeID, ctx: ZwibblerContext, e: Event): void;
    image?: string;
    hoverImage?: string;
}

export interface EColour {
    "r": number,
    "g": number,
    "b": number,
    "a": number
}